representative_signvectors fix for small cardinalities
Added a 0-length character value for representative_signvectors[[1]] to make it match the type of higher list entries. (The length really should be 0 because there are no relevant comparisons for a single note.)

For representative_signvectors[[2]], the sign vectors are individual numbers -- the old file had these mistakenly combined into a single entry "-1, 1". I've now split them into "-1" and "1" separately, so colornum should give correct results when called on dyads.